+++ deepinv/physics/__init__.py
@@ -29,3 +29,6 @@ from .phase_retrieval import (
     RandomPhaseRetrieval,
 )
 from .radio import RadioInterferometry
+
+from . import generator
+from . import functional