+++ deepinv/physics/generator/inpainting.py
@@ -1,9 +1,8 @@
 from typing import Tuple, Union
 from warnings import warn
-
-import numpy as np
 import torch
 from deepinv.physics.generator import PhysicsGenerator
+from deepinv.physics.functional import random_choice
 
 
 class BernoulliSplittingMaskGenerator(PhysicsGenerator):
@@ -46,18 +45,19 @@ class BernoulliSplittingMaskGenerator(PhysicsGenerator):
         split_ratio: float,
         pixelwise: bool = True,
         device: torch.device = torch.device("cpu"),
+        dtype: torch.dtype = torch.float32,
         rng: torch.Generator = None,
         *args,
         **kwargs,
     ):
-        super().__init__(*args, **kwargs)
+        super().__init__(*args, device=device, dtype=dtype, rng=rng, **kwargs)
         self.tensor_size = tensor_size
         self.split_ratio = split_ratio
         self.pixelwise = pixelwise
-        self.device = device
-        self.rng = rng or torch.Generator(device=self.device).manual_seed(0)
 
-    def step(self, batch_size=1, input_mask: torch.Tensor = None, **kwargs) -> dict:
+    def step(
+        self, batch_size=1, input_mask: torch.Tensor = None, seed: int = None, **kwargs
+    ) -> dict:
         r"""
         Generate a random mask.
 
@@ -66,12 +66,17 @@ class BernoulliSplittingMaskGenerator(PhysicsGenerator):
 
         :param int batch_size: batch_size. If None, no batch dimension is created. If input_mask passed and has its own batch dimension > 1, batch_size is ignored.
         :param torch.Tensor, None input_mask: optional mask to be split. If None, all pixels are considered. If not None, only pixels where mask==1 are considered. input_mask shape can optionally include a batch dimension.
+        :param int seed: the seed for the random number generator.
+
         :return: dictionary with key **'mask'**: tensor of size ``(batch_size, *tensor_size)`` with values in {0, 1}.
         :rtype: dict
         """
+        self.rng_manual_seed(seed)
+
         if isinstance(input_mask, torch.Tensor) and len(input_mask.shape) > len(
             self.tensor_size
         ):
+            input_mask = input_mask.to(self.device)
             if input_mask.shape[0] > 1:
                 # Batch dim exists in input_mask and it's > 1
                 batch_size = input_mask.shape[0]
@@ -98,7 +103,7 @@ class BernoulliSplittingMaskGenerator(PhysicsGenerator):
         return {"mask": mask}
 
     def check_pixelwise(self, input_mask=None) -> bool:
-        """Check if pixelwise can be used given input_mask dimensions and tensor_size dimensions"""
+        r"""Check if pixelwise can be used given input_mask dimensions and tensor_size dimensions"""
         pixelwise = self.pixelwise
 
         if pixelwise and len(self.tensor_size) == 2:
@@ -141,6 +146,7 @@ class BernoulliSplittingMaskGenerator(PhysicsGenerator):
         pixelwise = self.check_pixelwise(input_mask)
 
         if isinstance(input_mask, torch.Tensor) and input_mask.numel() > 1:
+            input_mask = input_mask.to(self.device)
             # Sample indices from given input mask
             if pixelwise:
                 idx = input_mask[0, ...].nonzero(as_tuple=False)
@@ -204,7 +210,7 @@ class GaussianSplittingMaskGenerator(BernoulliSplittingMaskGenerator):
     :param float std_scale: scale parameter of 2D Gaussian, in pixels.
     :param int, tuple[int] center_block: size of block in image center that is always kept for MRI autocalibration signal. Either int for square block or 2-tuple (h, w)
     :param str, torch.device device: device where the tensor is stored (default: 'cpu').
-    :param np.random.Generator rng: numpy random number generator. Note this is different from :class:`deepinv.physics.generator.BernoulliSplittingMaskGenerator` which requires torch generator.
+    :param torch.Generator rng: random number generator.
     """
 
     def __init__(
@@ -215,7 +221,7 @@ class GaussianSplittingMaskGenerator(BernoulliSplittingMaskGenerator):
         std_scale: float = 4.0,
         center_block: Union[Tuple[int], int] = (8, 8),
         device: torch.device = torch.device("cpu"),
-        rng: np.random.Generator = None,
+        rng: torch.Generator = None,
         *args,
         **kwargs,
     ):
@@ -238,7 +244,6 @@ class GaussianSplittingMaskGenerator(BernoulliSplittingMaskGenerator):
             if isinstance(center_block, int)
             else center_block
         )
-        self.rng = rng or np.random.default_rng(0)
 
     def batch_step(self, input_mask: torch.Tensor = None) -> dict:
         r"""
@@ -275,18 +280,20 @@ class GaussianSplittingMaskGenerator(BernoulliSplittingMaskGenerator):
         nx, ny = input_mask.shape[-2:]
         centerx, centery = nx // 2, ny // 2
 
-        x, y = np.meshgrid(np.arange(0, nx, 1), np.arange(0, ny, 1), indexing="ij")
+        x, y = torch.meshgrid(
+            torch.arange(0, nx, 1, device=self.device),
+            torch.arange(0, ny, 1, device=self.device),
+            indexing="ij",
+        )
 
         # Create PDF
-        gaussian = np.exp(
+        gaussian = torch.exp(
             -(
                 (x - centerx) ** 2 / (2 * (nx / self.std_scale) ** 2)
                 + (y - centery) ** 2 / (2 * (ny / self.std_scale) ** 2)
             )
         )
-        prob_mask = input_mask * torch.tensor(
-            gaussian[..., :, :], device=input_mask.device
-        )
+        prob_mask = input_mask * gaussian[..., :, :]
 
         prob_mask[
             ...,
@@ -301,14 +308,15 @@ class GaussianSplittingMaskGenerator(BernoulliSplittingMaskGenerator):
 
         for c in range(_C):
             for t in range(_T):
-                ind = self.rng.choice(
+                ind = random_choice(
                     nx * ny,
                     size=(input_mask[c, t, :, :].sum() * (1 - self.split_ratio))
                     .ceil()
                     .int()
                     .item(),
-                    p=norm_prob[c, t, :, :].cpu().flatten(),
+                    p=norm_prob[c, t, :, :].flatten(),
                     replace=False,
+                    rng=self.rng,
                 )
                 mask_out[c, t, ind] = 1
 
@@ -396,7 +404,7 @@ class Artifact2ArtifactSplittingMaskGenerator(Phase2PhaseSplittingMaskGenerator)
         device: torch.device = "cpu",
         rng: torch.Generator = None,
     ):
-        super().__init__(tensor_size, device, rng)
+        super().__init__(tensor_size, device, rng=rng)
         self.split_size = split_size
         self.prev_idx = None
         self.prev_split_size = None