+++ deepinv/physics/lidar.py
@@ -26,9 +26,10 @@ class SinglePhotonLidar(Physics):
     :param float sigma: Standard deviation of the Gaussian impulse response function.
     :param int bins: Number of histogram bins per pixel.
     :param str device: Device to use (gpu or cpu).
+    :param torch.Generator (Optional) rng: a pseudorandom random number generator for the Poisson noise model :meth:`deepinv.physics.noise.PoissonNoise`
     """
 
-    def __init__(self, sigma=1.0, bins=50, device="cpu"):
+    def __init__(self, sigma=1.0, bins=50, device="cpu", rng: torch.Generator = None):
         super().__init__()
 
         self.T = bins
@@ -36,7 +37,7 @@ class SinglePhotonLidar(Physics):
         self.sigma = torch.nn.Parameter(
             torch.tensor(sigma, device=device), requires_grad=False
         )
-        self.noise_model = PoissonNoise()
+        self.noise_model = PoissonNoise(rng=rng)
 
         h = ((self.grid - 3 * sigma) / self.sigma).pow(2)
         h = torch.exp(-h / 2.0)