+++ deepinv/physics/noise.py
@@ -1,7 +1,93 @@
 import torch
+import torch.nn as nn
+from typing import Callable
+import warnings
 
 
-class GaussianNoise(torch.nn.Module):
+class NoiseModel(nn.Module):
+    r"""
+
+    Base class for noise  model.
+    NoiseModel can be combined via :meth:`deepinv.physics.noise.NoiseModel.__mul__`,
+
+    :param torch.Generator (Optional) rng: a pseudorandom random number generator for the parameter generation. If is provided, it should be on the same device as the input.
+    """
+
+    def __init__(self, noise_model: Callable = None, rng: torch.Generator = None):
+        super().__init__()
+        if noise_model is None:
+            noise_model = lambda x: x
+        self.noise_model = noise_model
+        self.rng = rng
+        if rng is not None:
+            self.initial_random_state = rng.get_state()
+
+    def forward(self, input: torch.Tensor, seed: int = None) -> torch.Tensor:
+        r"""
+        Add noise to the input
+        :param torch.Tensor input: input tensor
+        :param int seed: the seed for the random number generator.
+        """
+        self.rng_manual_seed(seed)
+        return self.noise_model(input)
+
+    def __mul__(self, other):
+        r"""
+        Concatenates two noise :math:`N = N_1 \circ N_2` via the mul operation
+
+        The resulting operator will add the noise from both noise models and keep the `rng` of :math:`N_1`.
+
+        :param deepinv.physics.noise.NoiseModel other: Physics operator :math:`A_2`
+        :return: (deepinv.physics.noise.NoiseModel) concatenated operator
+
+        """
+        noise_model = lambda x: self.noise_model(other.noise_model(x))
+        return NoiseModel(
+            noise_model=noise_model,
+            rng=self.rng,
+        )
+
+    def rng_manual_seed(self, seed: int = None):
+        r"""
+        Sets the seed for the random number generator.
+
+        :param int seed: the seed to set for the random number generator. If not provided, the current state of the random number generator is used.
+            Note: it will be ignored if the random number generator is not initialized.
+        """
+        if seed is not None:
+            if self.rng is not None:
+                self.rng = self.rng.manual_seed(seed)
+            else:
+                warnings.warn(
+                    "Cannot set seed for random number generator because it is not initialized. The `seed` parameter is ignored."
+                )
+
+    def reset_rng(self):
+        r"""
+        Reset the random number generator to its initial state.
+        """
+        self.rng.set_state(self.initial_random_state)
+
+    def rand_like(self, input: torch.Tensor, seed: int = None):
+        r"""
+        Equivalent to `torch.rand_like` but supports a pseudorandom number generator argument.
+        :param int seed: the seed for the random number generator, if `rng` is provided.
+
+        """
+        self.rng_manual_seed(seed)
+        return torch.empty_like(input).uniform_(generator=self.rng)
+
+    def randn_like(self, input: torch.Tensor, seed: int = None):
+        r"""
+        Equivalent to `torch.randn_like` but supports a pseudorandom number generator argument.
+        :param int seed: the seed for the random number generator, if `rng` is provided.
+
+        """
+        self.rng_manual_seed(seed)
+        return torch.empty_like(input).normal_(generator=self.rng)
+
+
+class GaussianNoise(NoiseModel):
     r"""
 
     Gaussian noise :math:`y=z+\epsilon` where :math:`\epsilon\sim \mathcal{N}(0,I\sigma^2)`.
@@ -21,20 +107,22 @@ class GaussianNoise(torch.nn.Module):
         >>> y = physics(x)
 
     :param float sigma: Standard deviation of the noise.
-
+    :param torch.Generator (Optional) rng: a pseudorandom random number generator for the parameter generation.
     """
 
-    def __init__(self, sigma=0.1):
-        super().__init__()
+    def __init__(self, sigma=0.1, rng: torch.Generator = None):
+        super().__init__(rng=rng)
         self.update_parameters(sigma)
 
-    def forward(self, x, sigma=None, **kwargs):
+    def forward(self, x, sigma=None, seed=None, **kwargs):
         r"""
         Adds the noise to measurements x
 
         :param torch.Tensor x: measurements
         :param float, torch.Tensor sigma: standard deviation of the noise.
             If not None, it will overwrite the current noise level.
+        :param int seed: the seed for the random number generator, if `rng` is provided.
+
         :returns: noisy measurements
         """
         self.update_parameters(sigma)
@@ -43,8 +131,7 @@ class GaussianNoise(torch.nn.Module):
             sigma = sigma[(...,) + (None,) * (x.dim() - 1)]
         else:
             sigma = self.sigma
-
-        return x + torch.randn_like(x) * sigma
+        return x + self.randn_like(x, seed=seed) * sigma
 
     def update_parameters(self, sigma=None, **kwargs):
         r"""
@@ -56,7 +143,7 @@ class GaussianNoise(torch.nn.Module):
             self.sigma = to_nn_parameter(sigma)
 
 
-class UniformGaussianNoise(torch.nn.Module):
+class UniformGaussianNoise(NoiseModel):
     r"""
     Gaussian noise :math:`y=z+\epsilon` where
     :math:`\epsilon\sim \mathcal{N}(0,I\sigma^2)` and
@@ -79,32 +166,40 @@ class UniformGaussianNoise(torch.nn.Module):
 
     :param float sigma_min: minimum standard deviation of the noise.
     :param float sigma_max: maximum standard deviation of the noise.
+    :param torch.Generator (Optional) rng: a pseudorandom random number generator for the parameter generation.
 
     """
 
-    def __init__(self, sigma_min=0.0, sigma_max=0.5):
-        super().__init__()
+    def __init__(self, sigma_min=0.0, sigma_max=0.5, rng: torch.Generator = None):
+        super().__init__(rng=rng)
         self.sigma_min = sigma_min
         self.sigma_max = sigma_max
 
-    def forward(self, x, **kwargs):
+    def forward(self, x, seed: int = None, **kwargs):
         r"""
         Adds the noise to measurements x.
 
         :param torch.Tensor x: measurements.
+        :param int seed: the seed for the random number generator, if `rng` is provided.
+
         :returns: noisy measurements.
         """
-
+        self.rng_manual_seed(seed)
         sigma = (
-            torch.rand((x.shape[0], 1) + (1,) * (x.dim() - 2), device=x.device)
+            torch.rand(
+                (x.shape[0], 1) + (1,) * (x.dim() - 2),
+                device=x.device,
+                dtype=x.dtype,
+                generator=self.rng,
+            )
             * (self.sigma_max - self.sigma_min)
             + self.sigma_min
         )
-        noise = torch.randn_like(x) * sigma
+        noise = self.randn_like(x) * sigma
         return x + noise
 
 
-class PoissonNoise(torch.nn.Module):
+class PoissonNoise(NoiseModel):
     r"""
 
     Poisson noise :math:`y = \mathcal{P}(\frac{x}{\gamma})`
@@ -129,28 +224,33 @@ class PoissonNoise(torch.nn.Module):
     :param float gain: gain of the noise.
     :param bool normalize: normalize the output.
     :param bool clip_positive: clip the input to be positive before adding noise. This may be needed when a NN outputs negative values e.g. when using LeakyReLU.
+    :param torch.Generator (Optional) rng: a pseudorandom random number generator for the parameter generation.
 
     """
 
-    def __init__(self, gain=1.0, normalize=True, clip_positive=False):
-        super().__init__()
+    def __init__(
+        self, gain=1.0, normalize=True, clip_positive=False, rng: torch.Generator = None
+    ):
+        super().__init__(rng=rng)
         self.normalize = to_nn_parameter(normalize)
         self.update_parameters(gain)
         self.clip_positive = clip_positive
 
-    def forward(self, x, gain=None, **kwargs):
+    def forward(self, x, gain=None, seed: int = None, **kwargs):
         r"""
         Adds the noise to measurements x
 
         :param torch.Tensor x: measurements
         :param None, float, torch.Tensor gain: gain of the noise. If not None, it will overwrite the current noise level.
+        :param int seed: the seed for the random number generator, if `rng` is provided.
 
         :returns: noisy measurements
         """
         self.update_parameters(gain)
-
+        self.rng_manual_seed(seed)
         y = torch.poisson(
-            torch.clip(x / self.gain, min=0.0) if self.clip_positive else x / self.gain
+            torch.clip(x / self.gain, min=0.0) if self.clip_positive else x / self.gain,
+            generator=self.rng,
         )
         if self.normalize:
             y *= self.gain
@@ -166,7 +266,7 @@ class PoissonNoise(torch.nn.Module):
             self.gain = to_nn_parameter(gain)
 
 
-class GammaNoise(torch.nn.Module):
+class GammaNoise(NoiseModel):
     r"""
     Gamma noise :math:`y = \mathcal{G}(\ell, x/\ell)`
 
@@ -177,11 +277,13 @@ class GammaNoise(torch.nn.Module):
     Distribution for modelling speckle noise (eg. SAR images),
     where :math:`\ell>0` controls the noise level (smaller values correspond to higher noise).
 
+    .. warning:: This noise model does not support the random number generator.
+
     :param float, torch.Tensor l: noise level.
     """
 
     def __init__(self, l=1.0):
-        super().__init__()
+        super().__init__(rng=None)
         if isinstance(l, int):
             l = float(l)
         self.update_parameters(l)
@@ -208,7 +310,7 @@ class GammaNoise(torch.nn.Module):
             self.l = to_nn_parameter(l)
 
 
-class PoissonGaussianNoise(torch.nn.Module):
+class PoissonGaussianNoise(NoiseModel):
     r"""
     Poisson-Gaussian noise :math:`y = \gamma z + \epsilon` where :math:`z\sim\mathcal{P}(\frac{x}{\gamma})`
     and :math:`\epsilon\sim\mathcal{N}(0, I \sigma^2)`.
@@ -229,14 +331,14 @@ class PoissonGaussianNoise(torch.nn.Module):
 
     :param float gain: gain of the noise.
     :param float sigma: Standard deviation of the noise.
-
+    :param torch.Generator (Optional) rng: a pseudorandom random number generator for the parameter generation.
     """
 
-    def __init__(self, gain=1.0, sigma=0.1):
-        super().__init__()
+    def __init__(self, gain=1.0, sigma=0.1, rng: torch.Generator = None):
+        super().__init__(rng=rng)
         self.update_parameters(gain, sigma)
 
-    def forward(self, x, gain=None, sigma=None, **kwargs):
+    def forward(self, x, gain=None, sigma=None, seed: int = None, **kwargs):
         r"""
         Adds the noise to measurements x
 
@@ -244,13 +346,14 @@ class PoissonGaussianNoise(torch.nn.Module):
         :param None, float, torch.Tensor gain: gain of the noise. If not None, it will overwrite the current gain.
         :param None, float, torch.Tensor sigma: Tensor containing gain and standard deviation.
             If not None, it will overwrite the current gain and standard deviation.
+        :param int seed: the seed for the random number generator, if `rng` is provided.
+
         :returns: noisy measurements
         """
         self.update_parameters(gain, sigma)
-
-        y = torch.poisson(x / self.gain) * self.gain
-
-        y += torch.randn_like(x) * self.sigma
+        self.rng_manual_seed(seed)
+        y = torch.poisson(x / self.gain, generator=self.rng) * self.gain
+        y += self.randn_like(x) * self.sigma
         return y
 
     def update_parameters(self, gain=None, sigma=None, **kwargs):
@@ -267,7 +370,7 @@ class PoissonGaussianNoise(torch.nn.Module):
             self.sigma = to_nn_parameter(sigma)
 
 
-class UniformNoise(torch.nn.Module):
+class UniformNoise(NoiseModel):
     r"""
     Uniform noise :math:`y = x + \epsilon` where :math:`\epsilon\sim\mathcal{U}(-a,a)`.
 
@@ -286,23 +389,24 @@ class UniformNoise(torch.nn.Module):
         >>> y = physics(x)
 
     :param float a: amplitude of the noise.
+    :param torch.Generator (Optional) rng: a pseudorandom random number generator for the parameter generation.
     """
 
-    def __init__(self, a=0.1):
-        super().__init__()
+    def __init__(self, a=0.1, rng: torch.Generator = None):
+        super().__init__(rng=rng)
         self.update_parameters(a)
 
-    def forward(self, x, a=None, **kwargs):
+    def forward(self, x, a=None, seed: int = None, **kwargs):
         r"""
         Adds the noise to measurements x
 
         :param torch.Tensor x: measurements
         :param float, torch.Tensor a: amplitude of the noise. If not None, it will overwrite the current noise level.
+        :param int seed: the seed for the random number generator, if `rng` is provided.
         :returns: noisy measurements
         """
         self.update_parameters(a)
-
-        return x + (torch.rand_like(x) - 0.5) * 2 * self.a
+        return x + (self.rand_like(x, seed=seed) - 0.5) * 2 * self.a
 
     def update_parameters(self, a=None, **kwargs):
         r"""
@@ -314,7 +418,7 @@ class UniformNoise(torch.nn.Module):
             self.a = to_nn_parameter(a)
 
 
-class LogPoissonNoise(torch.nn.Module):
+class LogPoissonNoise(NoiseModel):
     r"""
     Log-Poisson noise :math:`y = \frac{1}{\mu} \log(\frac{\mathcal{P}(\exp(-\mu x) N_0)}{N_0})`.
 
@@ -344,13 +448,14 @@ class LogPoissonNoise(torch.nn.Module):
 
 
     :param float mu: normalization constant
+    :param torch.Generator (Optional) rng: a pseudorandom random number generator for the parameter generation.
     """
 
-    def __init__(self, N0=1024.0, mu=1 / 50.0):
-        super().__init__()
+    def __init__(self, N0=1024.0, mu=1 / 50.0, rng: torch.Generator = None):
+        super().__init__(rng=rng)
         self.update_parameters(mu, N0)
 
-    def forward(self, x, mu=None, N0=None, **kwargs):
+    def forward(self, x, mu=None, N0=None, seed: int = None, **kwargs):
         r"""
         Adds the noise to measurements x
 
@@ -359,11 +464,12 @@ class LogPoissonNoise(torch.nn.Module):
             If not None, it will overwrite the current number of photons.
         :param None, float, torch.Tensor N0: normalization constant.
             If not None, it will overwrite the current normalization constant.
+        :param int seed: the seed for the random number generator, if `rng` is provided.
         :returns: noisy measurements
         """
         self.update_parameters(mu, N0)
-
-        N1_tilde = torch.poisson(self.N0 * torch.exp(-x * self.mu))
+        self.rng_manual_seed(seed)
+        N1_tilde = torch.poisson(self.N0 * torch.exp(-x * self.mu), generator=self.rng)
         y = -torch.log(N1_tilde / self.N0) / self.mu
         return y